Make radio browser read-only in frozen builds
The radio browser is a very powerful tool, but it is also very
dangerous, as it allows arbitrary tweaking of any bit in memory
without any checking or safeguards. It has become clear that a lot
of regular users "out there" have this enabled and have taken to
poking around in memory, which seems to have resulted in a number of
concerning cases recently. This makes the radio browser effectively
read-only in the bundled builds to prevent this sort of tampering.
Developers running from source will still be able to use this, of
course, along with any other user that wants it bad enough to set up
a development environment.

Fixes: #6325