bitwise: Avoid duplicate definitions on one scope
This makes bitwise warn about duplicate names defined in the same
scope, which before resulted in loss of access to the eclipsed
element. Now, an error log is emitted.

There are a lot of drivers committing this crime, so I've added a
test to highlight them and marked it as XFAIL for the moment.

Fixes #11140